Add asmcat, bincat
These tools are currently fairly useless for real-world programming. Each of
them takes in either asm or binary, translates this to the internal intermediate
representation of a list of instruction structs, and then translates and outputs
in the same format that they took it in. These are mostly oly useful for testing
purposes, but may one day be useful if e.g. an optimiser stage is put in between
the input and output stages.