lex: empty file is equivalent to EOL
lex() returning NULL is used to convey an error case but having not allocated
any tokens in the case of an empty input file isn't an error case. This patch
causes lex to treat an empty token stream after successfully examining a file
as just a single EOL token. This is a fair approximation of an empty file for
this assembler's purposes, and results in the correct behaviour of an empty
output file.